// convex/_generated/api.d.ts
export declare const components: {
  betterAuth: {
    adapter: {
      create: FunctionReference<
        "mutation",
        "internal",
        {
          input:
            | {
                data: {
// ...
                  userId?: null | string;
                  username?: null | string;
                };
// ...
              };
          onCreateHandle?: string;
          select?: Array<string>;
        },
// ...
    };
// ...
  };
};